Annotate data.frame

This function takes in a dataframe that is result of post_process_df (or track_components) and an integer t (total time points), and returns a dataframe with 4 columns: count, added, deleted, and survivor. count is the number of cells at time point t. added is the number of new cells in time point t (relative to time point t - 1); added is always NA for t = 1. deleted is the number of cells at t that do not show up at later time points. survivor is the number of cells at time point t that remain from time point 1.

Usage

annotate_df(cell_df, t)

Arguments

cell_df

A dataframe that is the result from pos

t

An integer value

Value

A dataframe with 4 columns: count, addition, deletion, and survivor.